\name{DecisTreePredact}
\alias{}
\title{DecisTreePredact}
\usage{
randem forest model
}
\description{
this Function will run the random forest model and return a plot of the importance of each variable in the model
 we will use a cross validation with 10 number of folds and times = 2
}
\examples{
DecisTreePredact(0.7)
}
