\name{LevelsIris}
\alias{}
\title{randomForest}
\usage{
randem forest model
}
\description{
this Function will run the random forest model and return a plot of the importance of each variable in the model
 we will use a cross validation with 10 number of folds and times = 2
}
\examples{
RndmForstIris(0.7)
}
